notes crud: validation and authentication

+++ src/services/notes.js
@@ -1,3 +1,4 @@
+const boom = require('@hapi/boom');
 const MongoLib = require('../lib/mongo');
 
 class NotesService {
@@ -6,14 +7,23 @@ class NotesService {
     this.mongoDB = new MongoLib();
   }
 
-  async getNotes() {
-    const notes = await this.mongoDB.getAll(this.collection);
+  async getNotes({ userId }) {
+    const query = userId && { userId };
+    const notes = await this.mongoDB.getAll(this.collection, query);
     return notes || [];
   }
 
-  async getNote({ id }) {
+  async getNote({ id, userId }) {
     const note = await this.mongoDB.get(this.collection, id);
-    return note || {};
+    if (note) {
+      if (note.userId.toString() === userId.toString()) {
+        return note;
+      } else {
+        throw boom.unauthorized();
+      }
+    } else {
+      return {};
+    }
   }
 
   async createNote({ data }) {
@@ -21,14 +31,32 @@ class NotesService {
     return createdNoteId;
   }
 
-  async updateNote({ id, data }) {
-    const updatedNoteId = await this.mongoDB.update(this.collection, id, data);
-    return updatedNoteId;
+  async updateNote({ id, data, userId }) {
+    try {
+      const foundedNote = await this.getNote({ id, userId });
+      if (foundedNote._id) {
+        const updatedNoteId = await this.mongoDB.update(this.collection, id, data);
+        return updatedNoteId;
+      } else {
+        throw boom.unauthorized();
+      }
+    } catch (error) {
+      throw error;
+    }
   }
 
-  async deleteNote({ id }) {
-    const deletedNoteId = await this.mongoDB.delete(this.collection, id);
-    return deletedNoteId;
+  async deleteNote({ id, userId }) {
+    try {
+      const foundedNote = await this.getNote({ id, userId });
+      if (foundedNote._id) {
+        const deletedNoteId = await this.mongoDB.delete(this.collection, id);
+        return deletedNoteId;
+      } else {
+        throw boom.unauthorized();
+      }
+    } catch (error) {
+      throw error;
+    }
   }
 }
